KAFKA-14462; [15/N] Make Result generic and rename it
